// app/types/PostDetail.ts
export type PostDetail = {
  id: string;
  title: string;
  updatedAt?: string;
  user: {
    email: string;
    id: string;
    image: string;
    name: string;
  };
  comments?: {
    createdAt: string;
    id: string;
    postId: string;
    message: string;
    userId: string;
    user: {
      email: string;
      id: string;
      image: string;
      name: string;
    };
  }[];
};